requirements.txt
asttokens==2.4.1
attrs==23.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
country-converter==1.2
coverage==7.5.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.51.0
greenlet==3.0.3
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kafka-python==2.0.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
pprintpp==0.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycountry==23.12.11
pycountry-convert==0.7.2
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.2
referencing==0.35.1
repoze.lru==0.7
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
